Move common GPIOs into gpio module; re-export specific ports to flatten namespace
